requirements.txt.txt
streamlit
pandas
pdfplumber
pytesseract
Pillow
plotly
rapidfuzz
st-gsheets-connection
xlsxwriter